Initial commit
Proxy DLL for sw_api.dll. Eventual goal is to replace the THQN server used by the game for matchmaking and tracking active games/lobbies. It goes down occasionally. Interrupts modding and game nights. Not a major issue yet since they usually fix it quickly. Regardless it'd be useful to have full control of this as we mod the game more deeply. It'll also be a nice learning project since I've never written a networking/game server from scratch.

This DLL will need to emulate several interfaces the vanilla DLL provides that themselves emulate different steam interfaces. This is how the GOG version operates on the same API without really using the steam API. I defined the skeletons for these interfaces. The interfaces and their VTables still need to be filled out, and we'll need to figure out what functions that game actually needs and implement those. It defintely uses one or more of the ISteamNetworking P2P functions.

The new DLL will also need a server to track active players and lobbies and act as a go-between for players trying to find and join matches. I'll probably implement that as a sub-project in this repo when I get to it.