More interface logging improvements
Converted type aliases to structs with the primitive as the base type. Needed to do this since the logger code generation was stripping the alias name from them as typedefs. Also needed to add IFormattable since the logger formatting was trying to print them as a general boxed object (since the logger takes a list of objects as the args param).