Make DLL backend swappable
Need to easily be able to swap between this DLL being a wrapper that logs IO from the original DLL, and being a full replacement. The full replacement (CommunityBackend) doesn't do yet. At we should define custom implementations for each of the ISteamXXX interfaces the GOG version of sw_api.dll implements, and return those + setup ISteamClient and CSteamAPI + call DynamicInit callback on first run. I'll do that in the next commit.