Basic layout
Trying a different design from the main branch. Attempting to put most of the networking logic in a C# client app so it's in a higher level language that's quicker and less error prone to develop with. C++ Shim dll will act as an interface between the game and the C# client. Will need implementation and testing to confirm that this is feasible.
- C++ shim DLL that will either log calls and pass them onto the original DLL, or pass the calls to the client
- C# client app that networking logic will be implemented in. Communicates with server.
- ASP.NET server that tracks active lobbies and matchmaking data.